Extract a separate UnitInfoForm component
